+++ NEWS.md
@@ -1,3 +1,10 @@
 # tetragon 1.0.0
 
 * Added a `NEWS.md` file to track changes to the package.
+
+# tetragon 1.1.0
+
+* Added pred score metrics.
+* Renewed error calculation with greybox package.
+* Removed correlation weighting.
+* Added empirical distribution for matrix expansion.